Enhances medicine details view with partial expansion support; adds i…

+++ lib/features/test_laboratory/data/models/get_similar_tests_response_model.dart
@@ -47,7 +47,7 @@ class SimilarTests {
   String testName;
   String code;
   String standardRate;
-  int writtenPercent;
+  double writtenPercent;
   String interpretation;
   String recommendation;
